Move let_unit_value to its own module
